feat(web): move Routine settings to Hub Settings
Routine module's Settings tab is now removed; all routine settings
(habits CRUD, tags, categories, archive, backup, calendar prefs) live
in Hub Settings > Рутина, matching the Fizruk/Finyk/Nutrition pattern.
The empty-state CTA in the calendar now opens the habit quick-add
dialog instead of switching to a settings tab.
