+++ apps/web/src/modules/routine/components/RoutineBottomNav.tsx
@@ -4,7 +4,7 @@ import {
   type ModuleBottomNavItem,
 } from "@shared/components/ui/ModuleBottomNav";
 
-export type RoutineMainTab = "calendar" | "stats" | "settings";
+export type RoutineMainTab = "calendar" | "stats";
 
 interface RoutineNavItem extends ModuleBottomNavItem {
   id: RoutineMainTab;
@@ -54,27 +54,6 @@ const NAV: readonly RoutineNavItem[] = [
         <line x1="6" y1="20" x2="6" y2="12" />
         <line x1="12" y1="20" x2="12" y2="4" />
         <line x1="18" y1="20" x2="18" y2="9" />
-      </svg>
-    ),
-  },
-  {
-    id: "settings",
-    label: "Налаштування",
-    panelId: "routine-panel-settings",
-    icon: (
-      <svg
-        width="22"
-        height="22"
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        strokeWidth="1.65"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        aria-hidden
-      >
-        <path d="M12.22 2h-.44a2 2 0 0 0-2 2v.18a2 2 0 0 1-1 1.73l-.43.25a2 2 0 0 1-2 0l-.15-.08a2 2 0 0 0-2.73.73l-.22.38a2 2 0 0 0 .73 2.73l.15.1a2 2 0 0 1 1 1.72v.51a2 2 0 0 1-1 1.74l-.15.09a2 2 0 0 0-.73 2.73l.22.38a2 2 0 0 0 2.73.73l.15-.08a2 2 0 0 1 2 0l.43.25a2 2 0 0 1 1 1.73V20a2 2 0 0 0 2 2h.44a2 2 0 0 0 2-2v-.18a2 2 0 0 1 1-1.73l.43-.25a2 2 0 0 1 2 0l.15.08a2 2 0 0 0 2.73-.73l.22-.39a2 2 0 0 0-.73-2.73l-.15-.08a2 2 0 0 1-1-1.74v-.5a2 2 0 0 1 1-1.74l.15-.09a2 2 0 0 0 .73-2.73l-.22-.38a2 2 0 0 0-2.73-.73l-.15.08a2 2 0 0 1-2 0l-.43-.25a2 2 0 0 1-1-1.73V4a2 2 0 0 0-2-2z" />
-        <circle cx="12" cy="12" r="3" />
       </svg>
     ),
   },
